+++ rdar/Recent/RecentViewController.swift
@@ -7,13 +7,16 @@
 //
 
 import UIKit
+import SafariServices
 
 class RecentViewController: UITableViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        navigationItem.title = "Recent"
+        navigationItem.title = "Recent".localized()
+
+        RadarURLOpener.shared.delegate = self
 
         // Uncomment the following line to preserve selection between presentations
         // self.clearsSelectionOnViewWillAppear = false
@@ -94,4 +97,41 @@ class RecentViewController: UITableViewController {
     }
     */
 
+}
+
+extension RecentViewController: RadarURLOpenerUI {
+    func openRadarInSafariViewController(_ radarID: RadarID, radarOption: RadarOption, readerMode: Bool) {
+        // Prepend radarID to self.array
+
+        let url = radarID.url(by: radarOption)
+        presentSafariViewController(url: url, readerMode: readerMode)
+    }
+
+    func ask(completion: @escaping (Result<BrowserOption>) -> Void) {
+
+    }
+
+    private func presentSafariViewController(url: URL, readerMode: Bool) {
+        let sfvc: SFSafariViewController = {
+            if #available(iOS 11.0, *) {
+                let config = SFSafariViewController.Configuration()
+                config.barCollapsingEnabled = false
+                config.entersReaderIfAvailable = readerMode
+
+                return SFSafariViewController(url: url, configuration: config)
+            } else {
+                return SFSafariViewController(url: url, entersReaderIfAvailable: readerMode)
+            }
+        }()
+
+        sfvc.preferredBarTintColor = .barTintColor
+
+        if let presented = self.presentedViewController {
+            presented.dismiss(animated: false) {
+                self.present(sfvc, animated: false, completion: nil)
+            }
+        } else {
+            self.present(sfvc, animated: false, completion: nil)
+        }
+    }
 }